requirements.txt
fastapi
uvicorn[standard]
gunicorn
pydantic
python-dotenv
line-bot-sdk
google-generativeai
gradio